DB3: product_packages.price_vnd int→bigint (chống tràn VND giá gói lớn)
int (~2,1 tỷ ₫) tràn thầm lặng với gói giá lớn / VND cao. Widen sang bigint,
đồng bộ với orders.amount_vnd (đã long từ AmountVndToBigint).

- ProductPackage.PriceVnd int→long + config không set column type nên chỉ
  đổi CLR type → migration AlterColumn<long> price_vnd sang bigint (reversible).
- DTO PriceVnd (CreatePackageRequest/UpdatePackageRequest/PackageResponse)
  int→long để giá lớn không bị cắt ở biên API (PackageResponse.ToResponse
  narrowing sẽ vỡ build nếu giữ int). PayOS SDK Amount/Price nhận long — 0 cast.
- WebhookServiceTests: bỏ cast (int) thừa (helper priceVnd đã long).

Payment test 117 → 118 (+1: round-trip PriceVnd 3.000.000.000 > int.MaxValue).

+++ src/services/Isas.PaymentService.Tests/WebhookServiceTests.cs
@@ -27,7 +27,7 @@ public class WebhookServiceTests
             Id = Guid.NewGuid(),
             Name = $"Pack {credits}",
             Type = PackageType.OneTime,
-            PriceVnd = (int)priceVnd,
+            PriceVnd = priceVnd,
             InterviewCredits = credits,
             IsActive = true,
             CreatedAt = DateTime.UtcNow
